+++ examples/profileInfo.cpp
@@ -8,6 +8,8 @@
 #include "../src/encryption/SelfVerificationStatus.h"
 #include "../src/encryption/DeviceVerificationFlow.h"
 #include "../src/encryption/VerificationManager.h"
+#include "../src/voip/InputDevices.h"
+#include "../src/voip/AudioInputControl.h"
 
 UserInformation loginInfo;
 QMap <QString, Timeline*> timelineMap;
@@ -143,9 +145,19 @@ int main(int argc, char *argv[]){
     });
     client->enableLogger(true,true);
     client->start();
+    // -----------------------------------------------------
     // login();
-    
-    // QThread::sleep(10);
+    // -----------------------------------------------------  Volume Control
+    // AudioInputControl *aic = new AudioInputControl(); 
+    // QTimer::singleShot(5000, [aic] {
+    //     qDebug() << "Current Volume Value:" << aic->getVolume("Built-in Audio Analog Stereo");
+    //     aic->setVolume("Built-in Audio Analog Stereo" ,0.25);
+    //     aic->deviceChanged("Built-in Audio Analog Stereo");           
+    // });
+    // QObject::connect(aic, &AudioInputControl::levelChanged, [](qreal level){
+    //     qDebug() << level;
+    // });
+    // -----------------------------------------------------  
     // authentication->logout();
     return app.exec();     
 }