+++ src/Application.h
@@ -1,1 +1,1 @@
-#define VERSION_LIBRARY "0.1.11"
+#define VERSION_LIBRARY "0.1.13"